+++ src/components/atoms/buttons/LinkDropdownButton.vue
@@ -1,25 +1,29 @@
 <template>
-  <ButtonBase
+  <RouterLink
+    :to="to"
     class="border-l-2 flex justify-between w-full items-center"
     :class="statusClass?.button"
   >
     <component :is="statusClass?.component" class="text-left" :class="statusClass?.text">
       <slot></slot>
     </component>
-    <IconBase icon="ChevronIcon" :size="16" class="text-fontSecondary hover:text-fontPrimary" />
-  </ButtonBase>
+  </RouterLink>
 </template>
 
 <script setup lang="ts">
 import { computed } from 'vue';
+import { useRoute } from 'vue-router';
 
-import ButtonBase from '@/components/atoms/buttons/ButtonBase.vue';
 import SmallSemiboldTypography from '@/components/atoms/typography/SmallSemiboldTypography.vue';
 import SmallTypography from '@/components/atoms/typography/SmallTypography.vue';
 
 const props = defineProps<{
-  active: boolean;
+  to: string;
 }>();
+
+const route = useRoute();
+
+const isActive = computed(() => route.path === props.to);
 
 const status = new Map([
   [
@@ -40,5 +44,5 @@ const status = new Map([
   ]
 ]);
 
-const statusClass = computed(() => status.get(props.active));
+const statusClass = computed(() => status.get(isActive.value));
 </script>