+++ src/stores/menu.ts
@@ -16,10 +16,12 @@ export const useMenuStore = defineStore('menu', () => {
           items: [
             {
               name: 'Reclutamiento',
+              to: '/',
               active: true // Aux prop to know which item is active
             },
             {
-              name: 'Candidatos',
+              name: 'OnBoarding',
+              to: '/on-boarding',
               active: false // Aux prop to know which item is active
             }
           ]